requirements.txt
torch 
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cu121
gymnasium
stable-baselines3[extra]
tensorboard
numpy
pillow
gradio
tqdm
requests